refactor tests outside of src dir
