+++ test/JsonFormatter.test.tsx
@@ -4,8 +4,8 @@ import jsonSchemaFieldSets from "../common/mockData/jsonSchemaFielSetMock.json";
 import expectedSchema from "../common/mockData/jsonSchemaExpectedMock.json";
 import expectedUISchema from "../common/mockData/uiSchemaExpectedMock.json";
 import expectedFieldSetUISchema from "../common/mockData/uiSchemaFielSetExpectedMock.json";
-import { validateJSONSchema } from "../validateJsonSchema";
-import { generateUISchema } from "../generateUISchema";
+import { validateJSONSchema } from "../src/validateJsonSchema";
+import { generateUISchema } from "../src/generateUISchema";
 import {
     JSON_SCHEMA_ID_$SCHEMA_FAKE_DATA,
     JSON_SCHEMA_EMPTY_CHOICES_FAKE_DATA,
